add gist-setup
