+++ todo.md
@@ -7,7 +7,8 @@ Step 1 of 4: change ownership while preserving behavior. Sync execution cleanup,
 ## Tasks
 
 - [x] Add `services/github-auth.ts` with ordinary functions for credential access, removal, PAT validation, credential readiness, and authenticated-client acquisition using `infrastructure/github/client.ts`. Keep credential persistence behind this boundary using the existing `githubPat` metadata key. Support the supplied PAT used by validation adapters as well as stored credentials, without persisting validation inputs or adding authentication requests to Gist operations. Add focused auth tests preserving blank-input handling, untrimmed tokens, and existing validation results/errors.
-- [ ] Extract destination configuration, client-consuming Gist validation, and Gist creation into `services/gist-setup.ts`; have ongoing sync in `services/github-sync.ts` acquire its client through auth and read destination configuration through setup. Migrate `entrypoints/background/message-handlers.ts` and other internal callers to the owning auth/setup/sync modules; remove superseded exports and forwarding adapters from `github-sync.ts`, including the `validatePat` re-export. Preserve the existing `GistSyncConfig`/RPC inputs and results used by the popup, with each handler wired to the owning services. Preserve PAT → Gist ID → enabled write order, partial-update defaults, validation with the supplied PAT, and creation's client → export → localization → request → destination/status persistence order. Keep the setup module independent of the sync service to avoid a cycle.
+- [x] Mechanically move `getGistSyncConfig`, `setGistSyncConfig`, `validateGistId`, and `createNewGist` from `services/github-sync.ts` into `services/gist-setup.ts`, keeping signatures and function bodies unchanged. Update caller and test imports directly, with no compatibility re-exports. Keep sync execution unchanged. Review this stage before proceeding to structural changes.
+- [ ] Restructure the extracted setup boundary: separate destination configuration from credential-inclusive configuration, make Gist validation and creation consume an authenticated client, and have ongoing sync acquire its client through auth and read destination configuration through setup. Preserve the existing `GistSyncConfig`/RPC inputs and results used by the popup, with each handler wired to the owning services. Preserve PAT → Gist ID → enabled write order, partial-update defaults, validation with the supplied PAT, and creation's client → export → localization → request → destination/status persistence order. Keep setup independent of sync and avoid compatibility exports. Add focused coverage for supplied clients without credential access and the changed boundaries.
 - [ ] Route alarm readiness in `entrypoints/background/index.ts` through auth readiness and setup configuration, preserving the enabled/credential/Gist-ID gate, startup wait, shared write queue, interval, and badge fallback. Add background regression coverage for ready, disabled, missing-credential, and missing-destination cases; keep the readiness check free of network requests.
 - [ ] Route credential preservation/restoration in `services/import-export.ts` through auth, retaining reset's credential-removal position and import's read → reset → truthy-credential restore sequence. Retain and adapt `services/__tests__/import-export.test.ts` and `services/__tests__/import-export-characterization.test.ts` to verify missing/empty PAT behavior, ignored imported credentials, partial failures, exact persistence order, and unchanged exported JSON.
 - [ ] Adapt `services/__tests__/github-sync.test.ts` and add focused setup and message-handler coverage for the extracted boundaries. Migrate test imports and mocks to the owning modules; verify existing RPC behavior without retaining the old service exports. Preserve existing checks for private Gist creation, exact payloads, push/pull/no-change decisions, missing configuration, error/status behavior, concurrency, timestamps, and partial failures. Verify setup can use a supplied client without credential access and each Gist operation retains one client across requests; reuse `test/utils/deferred.ts` where needed.
